requirements.txt
biopython==1.86
pydantic==2.12.5
pandas==3.0.1
numpy==2.4.1
dash==4.1.0
